feat(vestad): status.json + Status module as the banner's single source
Introduce a `status` module that owns `status.json` and the startup banner, so
`vestad status` renders the exact banner the daemon printed — never stale.

- `Status` (version/user/port/mode/expose_lan/lan_url/tunnel/pid) is built in
  memory at startup, persisted atomically to `status.json` before the API opens,
  and printed via `Status::print_banner`. The api key is never persisted; it is
  passed in, so no secret lands in the file.
- `vestad status` and the systemd-install path render the banner from
  `status.json`, gated on the persisted pid being live (`pid_is_live`), so a
  stale file from a stopped/previous daemon shows "not running" instead of a
  misleading box. The old `print_server_info`/`read_server_info`/`print_qr`
  display is removed; the systemd-install command just points to `vestad status`
  (no polling, no duplicated render).
- The tunnel supervisor keeps the tunnel field honest: `supervise_tunnel` takes a
  decoupled `Fn(bool)` callback and reports debounced edges — up on
  re-registration, down only after a sustained outage (TUNNEL_DOWN_SUSTAINED_SECS)
  so transient blips keep showing "enabled".
- Banner polish: api key on one line, a spacer before the QR, and the QR uses the
  lowest error-correction level for the smallest square, scannable code.
